Wrap some function pointers passed to MPI
Before this commit, when client code would pass a copy or delete
callback to MPI_Comm_create_keyval, if the callbacks had any MPI calls
inside them then the code would segfault.  The segfault was due to the
native MPI passing in the native MPI_Comm to the callback, and then
YogiMPI would segfault when tyring to covert the comm to a MPI_Comm
because Yogi did not know the comm was already converted.

This change captures the callback functions and wraps them in lambdas
that un-converts the comm back to a YogiMPI_Comm so it is safe for the
callbacks to contain MPI calls.

There are likely other failure modes associated with user-supplied
callbacks to MPI routines.  Hopefully this will be a good pattern to
repeat if we find similar issues with other callbacks.