Add a callDepth tracking mechanism.
This commit wraps almost every single native MPI call with a callDepth
increment/decrement.  Currently this is only used for indenting calls in
the yogimpi.log.* files.

There are a certain class of problems that are hard to diagnose, and the
commonality is that YogiMPI has called the native MPI and then that MPI
has called a callback that was wrapped by YogiMPI.  Adding this
callDepth and indenting the log file based on callDepth makes it much
easier to diagnose the problem.